fix/removeAppServiceProvider: Remove AppServiceProvider and move its contents to FilterServiceProvider
The AppServiceProvider has been removed, and its contents have been relocated to the FilterServiceProvider to streamline service provider management and improve the project structure.
