+++ composer.json
@@ -53,7 +53,6 @@
   "extra": {
     "laravel": {
       "providers": [
-        "Meius\\LaravelFilter\\Providers\\AppServiceProvider",
         "Meius\\LaravelFilter\\Providers\\FilterServiceProvider"
       ]
     }
