+++ src/Providers/FilterServiceProvider.php
@@ -5,6 +5,9 @@ declare(strict_types=1);
 namespace Meius\LaravelFilter\Providers;
 
 use Illuminate\Support\ServiceProvider;
+use Meius\LaravelFilter\Console\FilterCacheCommand;
+use Meius\LaravelFilter\Console\FilterClearCommand;
+use Meius\LaravelFilter\Console\FilterMakeCommand;
 use Meius\LaravelFilter\Services\Filter\FilterManager;
 
 /**
@@ -14,6 +17,27 @@ class FilterServiceProvider extends ServiceProvider
 {
     public function register(): void
     {
-        $this->app->singletonIf(FilterManager::class);
+        $this->mergeConfigFrom(
+            __DIR__ . '/../../config/filter.php',
+            'filter'
+        );
+
+        $this->app->singleton(FilterManager::class);
+    }
+
+
+    public function boot(): void
+    {
+        $this->publishes([
+            __DIR__ . '/../../config/filter.php' => $this->app->configPath('filter.php'),
+        ], 'config');
+
+        if ($this->app->runningInConsole()) {
+            $this->commands([
+                FilterMakeCommand::class,
+                FilterClearCommand::class,
+                FilterCacheCommand::class,
+            ]);
+        }
     }
 }
